+++ techpack/camera/drivers/oplus/cam_sensor_module/cam_ois/onsemi_fw/LC898128/PhoneUpdate.c
@@ -992,13 +992,13 @@ const UINT_8 PACT0Tbl[] = { 0xFF, 0xFF };	/* Dummy table */
 const UINT_8 PACT1Tbl[] = { 0x20, 0xDF };	/* [ACT_02][ACT_01][ACT_03][ACT_05] */
 
 
-UINT_8 SetAngleCorrection( float DegreeGap, UINT_8 SelectAct, UINT_8 Arrangement )
+UINT_8 SetAngleCorrection( long DegreeGap, UINT_8 SelectAct, UINT_8 Arrangement )
 {
-//	double OffsetAngle = 0.0f;
-//	double OffsetAngleV_slt = 0.0f;
-	double OffsetAngle = (double)0;
-	double OffsetAngleV_slt = (double)0;
-//	double OffsetAngleS_slt = 0.0f;
+//	long OffsetAngle = 0.0f;
+//	long OffsetAngleV_slt = 0.0f;
+	long OffsetAngle = (long)0;
+	long OffsetAngleV_slt = (long)0;
+//	long OffsetAngleS_slt = 0.0f;
 	INT_32 Slgx45x = 0, Slgx45y = 0;
 	INT_32 Slgy45y = 0, Slgy45x = 0;
 	INT_32 Slagx45x = 0, Slagx45y = 0;
@@ -1007,7 +1007,7 @@ UINT_8 SetAngleCorrection( float DegreeGap, UINT_8 SelectAct, UINT_8 Arrangement
 	UINT_8	UcCnvF = 0;
 
 //	if( ( DegreeGap > 180.0f) || ( DegreeGap < -180.0f ) ) return ( 1 );
-	if( ( DegreeGap > (float)180) || ( DegreeGap < (float)-180 ) ) return ( 1 );
+	if( ( DegreeGap > (long)180) || ( DegreeGap < (long)-180 ) ) return ( 1 );
 	if( Arrangement >= 2 ) return ( 1 );
 
 /************************************************************************/
@@ -1015,11 +1015,11 @@ UINT_8 SetAngleCorrection( float DegreeGap, UINT_8 SelectAct, UINT_8 Arrangement
 /************************************************************************/
 	switch(SelectAct) {
 //		case 0x00 :
-//			OffsetAngle = (double)( 45.0f + DegreeGap ) * 3.141592653589793238 / 180.0f ;
+//			OffsetAngle = (long)( 45.0f + DegreeGap ) * 3.141592653589793238 / 180.0f ;
 //			UcCnvF = PACT1Tbl[ Arrangement ];
 //			break;
 //		case 0x01 :
-//			OffsetAngle = (double)( DegreeGap ) * 3.141592653589793238 / 180.0f ;
+//			OffsetAngle = (long)( DegreeGap ) * 3.141592653589793238 / 180.0f ;
 //			UcCnvF = PACT1Tbl[ Arrangement ];
 //			break;
 //		case 0x02 :
@@ -1030,8 +1030,8 @@ UINT_8 SetAngleCorrection( float DegreeGap, UINT_8 SelectAct, UINT_8 Arrangement
 //		case 0x08 :
 //		case 0x09 :
 		default :
-//			OffsetAngle = (double)( DegreeGap ) * 3.141592653589793238 / 180.0f ;
-			OffsetAngle = (double)( DegreeGap ) * (double)77633864123 / (double)24711626453 / (double)180 ;
+//			OffsetAngle = (long)( DegreeGap ) * 3.141592653589793238 / 180.0f ;
+			OffsetAngle = (long)( DegreeGap ) * (long)77633864123 / (long)24711626453 / (long)180 ;
 			UcCnvF = PACT1Tbl[ Arrangement ];
 			break;
 //		default :
@@ -1059,11 +1059,11 @@ UINT_8 SetAngleCorrection( float DegreeGap, UINT_8 SelectAct, UINT_8 Arrangement
 	RamWrite32A( 0x864C , 		(UINT_32)Slgy45x );
 	
 	if(SelectAct == 0x00) {
-//		OffsetAngleV_slt = (double)( 45.0f ) * 3.141592653589793238 / 180.0f ;
-		OffsetAngleV_slt = (double)( 45 ) * (double)77633864123 / (double)24711626453 / (double)180 ;
+//		OffsetAngleV_slt = (long)( 45.0f ) * 3.141592653589793238 / 180.0f ;
+		OffsetAngleV_slt = (long)( 45 ) * (long)77633864123 / (long)24711626453 / (long)180 ;
 	}else{
-//		OffsetAngleV_slt = (double)( 0.0f ) * 3.141592653589793238 / 180.0f ;
-		OffsetAngleV_slt = (double)( 0 ) * (double)77633864123 / (double)24711626453 / (double)180 ;
+//		OffsetAngleV_slt = (long)( 0.0f ) * 3.141592653589793238 / 180.0f ;
+		OffsetAngleV_slt = (long)( 0 ) * (long)77633864123 / (long)24711626453 / (long)180 ;
 	}
 //	Slagx45x = (INT_32)( cos( OffsetAngleV_slt )*2147483647.0);
 //	Slagx45y = (INT_32)(-sin( OffsetAngleV_slt )*2147483647.0);
@@ -1074,7 +1074,7 @@ UINT_8 SetAngleCorrection( float DegreeGap, UINT_8 SelectAct, UINT_8 Arrangement
 	RamWrite32A( 0x86F0 , 			(UINT_32)Slagy45y );
 	RamWrite32A( 0x86F4 , 			(UINT_32)Slagy45x );
 
-//	OffsetAngleS_slt = (double)( -90.0f ) * 3.141592653589793238 / 180.0f ;
+//	OffsetAngleS_slt = (long)( -90.0f ) * 3.141592653589793238 / 180.0f ;
 //	Slagx45x = (INT_32)( cos( OffsetAngleS_slt )*2147483647.0);
 //	Slagx45y = (INT_32)(-sin( OffsetAngleS_slt )*2147483647.0);
 //	Slagy45y = (INT_32)( cos( OffsetAngleS_slt )*2147483647.0);
@@ -1282,7 +1282,7 @@ void	SscDis( void )
  #define		ACT_CHK_NUM		3756		
  #define		ACT_THR			0x000003E8	
 // #define		ACT_MARGIN		0.75f
-#define        ACT_MARGIN      ((float)75 / (float)100)
+#define        ACT_MARGIN      ((long)75 / (long)100)
  
 UINT_8	TstActMov( UINT_8 UcDirSel )
 {
@@ -1290,7 +1290,7 @@ UINT_8	TstActMov( UINT_8 UcDirSel )
 	INT_32	SlMeasureParameterNum ;
 	INT_32	SlMeasureParameterA , SlMeasureParameterB ;
 	UnllnVal	StMeasValueA  , StMeasValueB ;
-	float		SfLimit , Sfzoom , Sflenz , Sfshift ;
+	long		SfLimit , Sfzoom , Sflenz , Sfshift ;
 	UINT_32		UlLimit , Ulzoom , Ullenz , Ulshift , UlActChkLvl ;
 	UINT_8		i;
 	UINT_32		UlReturnVal;
@@ -1308,23 +1308,23 @@ UINT_8	TstActMov( UINT_8 UcDirSel )
 	}
 
 
-	SfLimit = (float)UlLimit / (float)0x7FFFFFFF;
+	SfLimit = (long)UlLimit / (long)0x7FFFFFFF;
 	if( Ulzoom == 0){
 		Sfzoom = 0;
 	}else{
-		Sfzoom = (float)abs(Ulzoom) / (float)0x7FFFFFFF;
+		Sfzoom = (long)abs(Ulzoom) / (long)0x7FFFFFFF;
 	}
 	if( Ullenz == 0){
 		Sflenz = 0;
 	}else{
-		Sflenz = (float)Ullenz / (float)0x7FFFFFFF;
+		Sflenz = (long)Ullenz / (long)0x7FFFFFFF;
 	}
 	Ulshift = ( Ulshift & 0x0000FF00) >> 8 ;	
 	Sfshift = 1;
 	for( i = 0 ; i < Ulshift ; i++ ){
 		Sfshift *= 2;
 	}
-	UlActChkLvl = (UINT_32)( (float)0x7FFFFFFF * SfLimit * Sfzoom * Sflenz * Sfshift * ACT_MARGIN );
+	UlActChkLvl = (UINT_32)( (long)0x7FFFFFFF * SfLimit * Sfzoom * Sflenz * Sfshift * ACT_MARGIN );
 
 	SlMeasureParameterNum	=	ACT_CHK_NUM ;
